Convert SingletonAnnotation to TS
Convert EagerSingletonAnnotation to TS
