+++ src/annotations/EagerSingletonAnnotation.ts
@@ -1,0 +1,5 @@
+import { SingletonAnnotation } from './SingletonAnnotation';
+
+export class EagerSingletonAnnotation extends SingletonAnnotation {
+  public static _name = 'EagerSingletonAnnotation';
+}
